# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.BadgeDefinition:
    properties:
      code:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
...
  domain.TokenPair:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  domain.UserBadge:
    properties:
      badge_id:
        type: integer
      challenge_id:
        type: string
      code:
        description: Joined fields
        type: string
      description:
        type: string
      earned_at:
        type: string
      icon:
        type: string
      id:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для веб-сервиса коллективного трекинга привычек
  title: Habix API
  version: "1.0"
paths:
  /auth/login:
    post:
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.TokenPair'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "204":
          description: No Content
      summary: Logout
      tags:
      - auth
  /auth/refresh:
    post:
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.TokenPair'
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.TokenPair'
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /badges:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.BadgeDefinition'
            type: array
      security:
      - BearerAuth: []
      summary: List all badge definitions
      tags:
      - badges
...
  /users/{id}/badges:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.UserBadge'
            type: array
      security:
      - BearerAuth: []
      summary: Get user's badges
      tags:
      - badges
...
  /users/me/badges:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.UserBadge'
            type: array
      security:
      - BearerAuth: []
      summary: Get current user's badges
      tags:
      - badges
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"